Created home layout
